Working with added functions
retr_datamock -> variables holding the return values now [this is a great function]
anet class -> trying to let each layer be added on in a modular format [ ie:
net.addFcon
net.addFcon
net.addFcon(whchLayr = 'last')
net.get_metr
# returns all our metrics]

genNet and gdatFile are both kinda achieving the same goal, but in slightly different ways:
genNet takes gdat as an argument and has a minimalist number of self.members 
gdatFile has all the initializing data as self.members and can be customized and modularized quite a bit with inputs. I took the 'here'/'ete6' input from the genNet and implemented it in the gdatFile... we'll see which we like more. (I'm using the gdatFile to try to modularize as much as possible, but it will likely be faster to get an immediate result from the genNet file? I honestly don't know at this point)